+++ echopype/process/process_azfp.py
@@ -68,9 +68,10 @@ class ProcessAZFP(ProcessBase):
         # Attached calculated range to the dataset
         if ed.range.ndim == 2:
             # Range does not have a ping time dimension when the temperature is a single value
-            Sv['range'] = (('frequency', 'ping_time'), ed.range)
+            Sv['range'] = (('frequency', 'range_bin'), ed.range.transpose('frequency', 'range_bin'))
         else:
-            Sv['range'] = (('frequency', 'ping_time', 'range_bin'), ed.range)
+            Sv['range'] = (('frequency', 'ping_time', 'range_bin'),
+                           ed.range.transpose('frequency', 'ping_time', 'range_bin'))
 
         # Save calibrated data into the calling instance and
         # to a separate .nc file in the same directory as the data
@@ -100,9 +101,10 @@ class ProcessAZFP(ProcessBase):
 
         # Attached calculated range to the dataset
         if ed.range.ndim == 2:
-            Sp['range'] = (('frequency', 'range_bin'), ed.range)
+            Sp['range'] = (('frequency', 'range_bin'), ed.range.transpose('frequency', 'range_bin'))
         else:
-            Sp['range'] = (('frequency', 'ping_time', 'range_bin'), ed.range)
+            Sp['range'] = (('frequency', 'ping_time', 'range_bin'),
+                           ed.range.transpose('frequency', 'ping_time', 'range_bin'))
 
         if save:
             # Update pointer in EchoData
