+++ echopype/utils/io.py
@@ -5,42 +5,6 @@ import os
 from collections.abc import MutableMapping
 
 
-# TODO: @ngkavin: if you are using this only for the process objects, this should not be here.
-#  My suggestion to factor this out was to have it work for both convert and process.
-#  Let's discuss what you are doing differently in both cases and see if can combine.
-
-def validate_proc_path(ed, postfix, save_path=None):
-    """Creates a directory if it doesn't exist. Returns a valid save path.
-    """
-    def _assemble_path():
-        file_in = os.path.basename(ed.raw_path[0])
-        file_name, file_ext = os.path.splitext(file_in)
-        return file_name + postfix + file_ext
-
-    if save_path is None:
-        save_dir = os.path.dirname(ed.raw_path[0])
-        file_out = _assemble_path()
-    else:
-        path_ext = os.path.splitext(save_path)[1]
-        # If given save_path is file, split into directory and file
-        if path_ext != '':
-            save_dir, file_out = os.path.split(save_path)
-            if save_dir == '':  # save_path is only a filename without directory
-                save_dir = os.path.dirname(ed.raw_path)  # use directory from input file
-        # If given save_path is a directory, get a filename from input .nc file
-        else:
-            save_dir = save_path
-            file_out = _assemble_path()
-
-    # Create folder if not already exists
-    if save_dir == '':
-        save_dir = os.getcwd()
-    if not os.path.exists(save_dir):
-        os.mkdir(save_dir)
-
-    return os.path.join(save_dir, file_out)
-
-
 def get_files_from_dir(folder):
     """Retrieves all Netcdf and Zarr files from a given folder"""
     valid_ext = ['.nc', '.zarr']
